Create file for events the dom
