+++ src/renderers/dom/render.js
@@ -1,4 +1,5 @@
 import { renderFactory } from '../../isomorphic/fiber';
+import { addEventListener, isEvent } from './events';
 
 /**
  * @public
@@ -47,37 +48,6 @@ const setAttributes = (component, node, attributes) => {
 
 		node.setAttribute(attr, attributes[attr]);
 	});
-}
-
-/**
- * @param component
- * @param node
- * @param attr
- * @param attributes
- * @internal
- */
-const addEventListener = (component, node, attr, attributes) => {
-	let eventName = extractEventName(attr);
-
-	node.addEventListener(eventName, attributes[attr].bind(component));
-}
-
-/**
- * @param attr
- * @return {boolean}
- * @internal
- */
-const isEvent = (attr) => {
-	return /^on/.test(attr);
-}
-
-/**
- * @param attr
- * @return {string}
- * @internal
- */
-const extractEventName = (attr) => {
-	return attr.slice('2').toLowerCase();
 }
 
 export default render;
